backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain==1.2.9",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.9",
    "langchain-google-genai==4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-mongodb==0.11.0",
    "langchain-qdrant==1.1.0",
    "langchain-tavily==0.2.17",
    "langgraph>=0.2.54",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.5.2",
    "livekit-agents[google]~=1.4",
    "livekit-plugins-deepgram~=1.4",
]

[tool.uv.workspace]
members = [
    "livekit-voice-agent",
]
